requirements.txt
ansible==2.1.1
netaddr>=0.7.18
gitpython>=0.6.4
PyYAML>=3.12
colorlog>=2.7.0
tabulate>=0.7.7

# support for ansible os_ modules
# keep that here to avoid duplication in plugins
openstacksdk==0.9.8
pytz>=2016.3
os-client-config==1.13.1
shade==1.7.0
requests>=2.9.1
Babel!=2.3.0,!=2.3.1,!=2.3.2,!=2.3.3,>=1.3
